requirements.txt
flask
flask_sqlalchemy
python-dotenv
mysql-connector-python
gunicorn
bcrypt
transformers
torch
tensorflow
pymysql
Flask-Bcrypt
mysqlclient
openai
pandas
scikit-learn
nltk
sentence-transformers
vaderSentiment
matplotlib
wordcloud


